add luamake script
