+++ .gitignore
@@ -1,2 +1,5 @@
 o
 *.dll
+*.so
+build
+.vscode
