+++ test.lua
@@ -102,170 +102,6 @@ do
 	print("log(v3, 3):", math3d.tostring(vv))
 end
 
-print "-----plane test-----"
-do
-	--[[
-		    local d1dotn = math3d.dot(plane, ray.d)
-			local t = 0.0
-			if math.abs(d1dotn) > 1e-7 then
-				local dis = math3d.index(plane, 4)
-				local odotn = math3d.dot(ray.o, plane)
-				t = (dis - odotn) / d1dotn
-			end
-			local intersetion_pt = math3d.muladd(ray.d, tt, ray.o)
-	]]
-	local plane_pos = math3d.vector(0, 3, 0)
-	local plane_dir = math3d.vector(0, 10, 0)
-	local plane = math3d.plane(plane_pos, plane_dir)
-
-	local ray = {o = math3d.vector(0, 10, 0), d = math3d.vector(0.0, -3, 0.0)}
-	local tt = math3d.plane_ray(ray.o, ray.d, plane)
-	if tt == nil then
-		print "plane parallel with ray"
-	elseif tt < 0 then
-		print "plane intersetion in ray backward"
-	else
-		print "plane intersetion with plane front face"
-	end
-
-	local intersetion_pt = math3d.muladd(ray.d, tt, ray.o)
-	print(math3d.tostring(intersetion_pt))
-end
-
-local corner_names = {
-	"lbn", "ltn", "rbn", "rtn",
-	"lbf", "ltf", "rbf", "rtf",
-}
-
-local function print_box_points(points, tabnum)
-	assert(math3d.array_size(points) == 8)
-
-	local function point_name(pidx, tabnum)
-		return ('\t'):rep(tabnum or 0) .. ("%s: %s"):format(corner_names[pidx], math3d.tostring(math3d.array_index(points, pidx)))
-	end
-
-	local function point_pairs(pidx1, pidx2, tabnum)
-		return ("%s;\t%s"):format(point_name(pidx1, tabnum), point_name(pidx2))
-	end
-	local t = {
-		point_pairs(1, 5, tabnum),
-		point_pairs(2, 6, tabnum),
-		point_pairs(3, 7, tabnum),
-		point_pairs(4, 8, tabnum),
-	}
-
-	return print(table.concat(t, "\n"))
-end
-
-local function print_box_planes(planes, tabnum)
-	assert(math3d.array_size(planes) == 6)
-
-	local function plane_name(pidx, tabnum)
-		return tu.tabs(tabnum) .. math3d.tostring(math3d.array_index(planes, pidx))
-	end
-	local t = {}
-	for i=1, 6 do
-		t[#t+1] = plane_name(i, tabnum)
-	end
-
-	print(table.concat(t, '\n'))
-end
-
-print "----- ray/line interset with triangles -----"
-do
-	local v0, v1, v2 = math3d.vector(-1.0, 0.0, 0.0, 1.0), math3d.vector(0.0, 1.0, 0.0, 1.0), math3d.vector(1.0, 0.0, 0.0, 1.0)
-	print "\ttriangle1:"
-	tu.print_triangle(v0, v1, v2, 1)
-
-	local v3, v4, v5 = math3d.vector(-1.0, 0.0, 2.0, 1.0), math3d.vector(0.0, 1.0, 2.0, 1.0), math3d.vector(1.0, 0.0, 2.0, 1.0)
-	print "\ttriangle2:"
-	tu.print_triangle(v3, v4, v5, 1)
-
-	local v6, v7, v8 = math3d.vector(0.0, 0.0, 1.0, 1.0), math3d.vector(0.0, 1.0, 0.0, 1.0), math3d.vector(0.0, 0.0, -1.0, 1.0)
-	print "\ttriangle3:"
-	tu.print_triangle(v6, v7, v8, 1)
-
-	local ray = {o=math3d.vector(0.0, 0.0, 0.0), d=math3d.vector(0.0, 0.0, 1.0)}
-	print "\tinterset with ray:"
-	tu.print_ray(ray, 1)
-
-	print "\tsegment1 interset with triangle:"
-	local s0, s1 = math3d.vector(0, 0, 1, 1), math3d.vector(0, 0, 3, 1)
-
-	print("\tstart point:", math3d.tostring(s0))
-	print("\tend point:", math3d.tostring(s1))
-
-	local interset, t = math3d.triangle_ray(ray.o, ray.d, v0, v1, v2)
-	if interset then
-		print(("\tray triangle interset result:%f, point:%s"):format(t, math3d.tostring(math3d.muladd(ray.d, t, ray.o))))
-	else
-		print "\tray NOT interset with triangle"
-	end
-
-	print"\tsegment with triangle1:"
-	interset, t = math3d.triangle_ray(s0, math3d.sub(s1, s0), v0, v1, v2)
-	print("\tsegment interset with triangle1: ", interset, "t: ", t, "is in the segment:", 0<=t and t<=1.0)
-
-	interset, t = math3d.triangle_ray(s0, math3d.sub(s1, s0), v3, v4, v5)
-	print("\tsegment interset with triangle2: ", interset, "t: ", t, "is in the segment:", 0<=t and t<=1.0)
-
-	interset, t = math3d.triangle_ray(s0, math3d.sub(s1, s0), v6, v7, v8)
-	print("\tsegment interset with triangle3: ", interset, "t: ", t or "no intersect")
-
-	print "\tray intersect with box:"
-	local boxcorners = {math3d.vector(-1.0, -1.0, -1.0), math3d.vector(1.0, 1.0, 1.0)}
-	local aabb = math3d.aabb(boxcorners[1], boxcorners[2])
-	local aabbpoints = math3d.aabb_points(aabb)
-
-	print "\tray intersect with aabb points:"
-	print_box_points(aabbpoints, 2)
-
-	local ray1 = {o=math3d.vector(0.0, 0.0, 0.0, 1.0), d=math3d.vector(0.0, 0.0, -1.0)}
-	local function do_ray_box_test(ray, boxpoints)
-		print "\tray:"
-		tu.print_ray(ray, 2)
-		local resultpoints = math3d.ray_intersect_box(ray.o, ray.d, boxpoints)
-		print "\tray intersect results:"
-		tu.print_points(resultpoints, 2)
-		return resultpoints
-	end
-
-	print "\tray box test 1:"
-	local results = do_ray_box_test(ray1, aabbpoints)
-	assert(math3d.array_size(results) == 1)
-
-	print "\tray box test 2:"
-	local ray2 = tu.create_ray(math3d.vector(0.0, 0.0, -1.0, 1.0), math3d.vector(-1.0, 0.0, 0.0, 1.0))
-	results = do_ray_box_test(ray2, aabbpoints)
-	assert(math3d.array_size(results) == 2)
-
-	print "\tray box test 3:"
-	--local ray3 = create_ray(math3d.array_index(aabbpoints, 1), math3d.array_index(aabbpoints, 5))
-	local ray3 = tu.create_ray(math3d.vector(0.0, 0.0, -1.0, 1.0), math3d.vector(0.0, 0.0, 1.0, 1.0))
-	results = do_ray_box_test(ray3, aabbpoints)
-	assert(math3d.array_size(results) == 2)
-
-	print "\tray box test 4:"
-	local ray4 = tu.create_ray(math3d.vector(-1.0, 0.0, 0.0, 1.0), math3d.vector(1.0, 0.0, 1.0, 1.0))
-	results = do_ray_box_test(ray4, aabbpoints)
-	assert(math3d.array_size(results) == 2)
-
-	local aabb2 = math3d.aabb(math3d.vector(0.0, 0.0, 0.0, 0.0), math3d.vector(2.0, 2.0, 2.0, 0.0))
-	local aabbpoints2 = math3d.aabb_points(aabb2)
-	print "\taabb2 points:"
-	print_box_points(aabbpoints2, 2)
-
-	print "\tray box test 5:"
-	local ray5 = tu.create_ray(math3d.vector(0.0, 0.0, 0.0, 1.0), math3d.vector(2.0, 2.0, 2.0, 1.0))
-	results = do_ray_box_test(ray5, aabbpoints2)
-	assert(math3d.array_size(results) == 2)
-
-	print "\tray box test 6:"
-	local ray6 = tu.create_ray(math3d.vector(1.0, 1.0, -1.0, 1.0), math3d.vector(2.0, 2.0, 2.0, 1.0))
-	results = do_ray_box_test(ray6, aabbpoints2)
-	assert(math3d.array_size(results) == 2)
-end
-
 print "===SRT==="
 do
 	ref1.m = { r = { 0, math.rad(60), 0 }, t = { 1,2,3} }	-- .s = 1
@@ -400,6 +236,8 @@ do
 	print("forward:", math3d.tostring(forward), "right:", math3d.tostring(right), "up:", math3d.tostring(up))
 end
 
+require "test.plane"
+require "test.raytriangle"
 require "test.frustum_aabb"
 require "test.array"
 require "test.matrix"
